Update camera.far and near from CSM.lightFar and lightNear
If I do not do any mistake, I think it was the initially purpose of lightFar and lightNear properties. But they aren't updated when the value changes. So to avoid hard update like for each lights do csm.light.shadow.camera.far = csm.lightFar it would be nice to already have it in csm.update.